- Upgraded lower supported XF version to 4.5 - Removed useless interface - Moved RangeSlider to Views namespace and folder
