+++ XamarinCommunityToolkitSample/Pages/Views/RangeSliderPage.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <pages:BasePage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:views="clr-namespace:Microsoft.Toolkit.Xamarin.Forms.UI.Controls;assembly=Microsoft.Toolkit.Xamarin.Forms"
+             xmlns:views="clr-namespace:Microsoft.Toolkit.Xamarin.Forms.UI.Views;assembly=Microsoft.Toolkit.Xamarin.Forms"
              xmlns:pages="clr-namespace:Microsoft.Toolkit.Xamarin.Forms.Sample.Pages"
              x:Class="Microsoft.Toolkit.Xamarin.Forms.Sample.Pages.Views.RangeSliderPage"
              xmlns:resources="clr-namespace:Microsoft.Toolkit.Xamarin.Forms.Sample.Resx">
